Phase 5: extract admin rescan form JS to external files
Move RescanFormController and streaming rescan progress handling from
admin/rescan.php into admin-rescan-form.js. Diff pane styles move to
admin-rescan.css.
