+++ AGENTS.md
@@ -92,14 +92,15 @@ the non-obvious steps to actually run the services.
 ### Security headers
 
 - `init.php` sends `Content-Security-Policy-Report-Only` alongside existing security
-  headers. Enforcement is deferred until remaining inline scripts (`admin/rescan.php`)
-  are extracted to external files.
+  headers. Public and admin page scripts are now externalized; CSP enforcement is the
+  next Phase 5 step.
 - Date localization uses `/js/localized-date-formatter.js`; changelog grouping uses
   `/js/changelog-date-grouping.js`; homepage queue polling uses
   `/js/player-queue-manager.js`; about-page scan log polling uses
   `/js/scan-log-renderer.js`; admin reported-player deletes use
   `/js/admin-report-delete.js`; admin log bulk actions use
-  `/js/admin-log-bulk-actions.js`; admin game merge UI uses `/js/admin-merge-form.js`.
+  `/js/admin-log-bulk-actions.js`; admin game merge UI uses `/js/admin-merge-form.js`;
+  admin game rescan UI uses `/js/admin-rescan-form.js`.
 
 ### Composer
